Move most of the code to src subfolder
