+++ src/jobs/packetgen.go
@@ -1,4 +1,4 @@
-package job
+package jobs
 
 import (
 	"context"
@@ -6,15 +6,17 @@ import (
 	"strconv"
 	"time"
 
-	"github.com/Arriven/db1000n/logs"
-	"github.com/Arriven/db1000n/metrics"
-	"github.com/Arriven/db1000n/packetgen"
-	"github.com/Arriven/db1000n/template"
 	"github.com/google/uuid"
+
+	"github.com/Arriven/db1000n/src/logs"
+	"github.com/Arriven/db1000n/src/metrics"
+	"github.com/Arriven/db1000n/src/packetgen"
+	"github.com/Arriven/db1000n/src/utils"
+	"github.com/Arriven/db1000n/src/utils/templates"
 )
 
 func packetgenJob(ctx context.Context, l *logs.Logger, args Args) error {
-	defer panicHandler()
+	defer utils.PanicHandler()
 	type packetgenJobConfig struct {
 		BasicJobConfig
 		Packet json.RawMessage
@@ -28,8 +30,8 @@ func packetgenJob(ctx context.Context, l *logs.Logger, args Args) error {
 		return err
 	}
 
-	host := template.Execute(jobConfig.Host)
-	port, err := strconv.Atoi(template.Execute(jobConfig.Port))
+	host := templates.Execute(jobConfig.Host)
+	port, err := strconv.Atoi(templates.Execute(jobConfig.Port))
 	if err != nil {
 		l.Error("error parsing port: %v", err)
 		return err
@@ -47,7 +49,7 @@ func packetgenJob(ctx context.Context, l *logs.Logger, args Args) error {
 		default:
 		}
 
-		packetConfigBytes := []byte(template.Execute(string(jobConfig.Packet)))
+		packetConfigBytes := []byte(templates.Execute(string(jobConfig.Packet)))
 		l.Debug("[packetgen] parsed packet config template:\n%s", string(packetConfigBytes))
 		var packetConfig packetgen.PacketConfig
 		err := json.Unmarshal(packetConfigBytes, &packetConfig)
