+++ src/synfloodraw/constants.go
