+++ src/synfloodraw/utils.go
